requirements.txt
# --- Web Framework ---
fastapi

# --- ASGI Server ---
uvicorn[standard]

# --- ORM e Banco de Dados ---
sqlmodel
alembic

# --- Driver do PostgreSQL ---
psycopg2-binary

# --- Gerenciamento de Configuração ---
pydantic-settings
python-dotenv

# --- Segurança ---
passlib[bcrypt]
